New Class
Separated menu to it’s own class, separate from main to execute program
